test invalid job-id for all controllers
